BYT-246 bug fix
